% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongituRF.R
\name{DataLongGenerator}
\alias{DataLongGenerator}
\title{Longitudinal data generator}
\usage{
DataLongGenerator(n = 50, p = 10, G = 10)
}
\arguments{
\item{n}{\link{numeric}: Number of individuals. The default value is \code{n=50}.}

\item{p}{\link{numeric}: Number of predictors. The default value is \code{p=10}.}

\item{G}{\link{numeric}: Number of groups of predictors. Default value is \code{G=6}}
}
\value{
a list of the following elements: \itemize{
\item \code{Y:} vector of the output trajectories.
\item \code{X :} matrix of the fixed-effects predictors.
\item \code{Z:} matrix of the random-effects predictors.
\item \code{id: } vector of the identifiers for each individual.
\item \code{time: } vector the the time measurements for each individual.
}
}
\description{
Longitudinal data generator
}
