% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongituRF.R
\name{Linmix}
\alias{Linmix}
\title{(S)LMEM algorithm}
\usage{
Linmix(X, Y, id, Z, iter, time, sto, delta = 0.001)
}
\arguments{
\item{X}{\link{matrix}:}

\item{Y}{\link{vector}:}

\item{id}{\link{vector}:}

\item{Z}{\link{matrix}:}

\item{iter}{\link{numeric}:}

\item{time}{\link{vector}:}

\item{sto}{\link{numeric}:}

\item{delta}{\link{numeric}:}
}
\value{

}
\description{
(S)LMEM algorithm
}
