% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongituRF.R
\name{MERF}
\alias{MERF}
\title{(S)MERF algorithm}
\usage{
MERF(X, Y, id, Z, iter, mtry, ntree, time, sto, delta = 0.001)
}
\arguments{
\item{X}{\link{matrix}:}

\item{Y}{\link{vector}:}

\item{id}{\link{vector}:}

\item{Z}{\link{matrix}:}

\item{iter}{\link{numeric}:}

\item{mtry}{\link{numeric}:}

\item{ntree}{\link{numeric}:}

\item{time}{\link{vector}:}

\item{sto}{\link{character}:}

\item{delta}{\link{numeric}:}
}
\value{

}
\description{
(S)MERF algorithm
}
