% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongituRF.R
\name{MERT}
\alias{MERT}
\title{(S)MERT algorithm}
\usage{
MERT(X, Y, id, Z, iter = 100, time, sto, delta = 0.001)
}
\arguments{
\item{X}{\link{matrix}: A \code{N}x\code{p} matrix containing the \code{p} predictors of the fixed effects, column codes for a predictor.}

\item{Y}{\link{vector}: A vector containing the output trajectories.}

\item{id}{\link{vector}: Is the vector of the identifiers for the different trajectories.}

\item{Z}{\link{matrix}: A \code{N}x\code{q} matrix containing the \code{q} predictor of the random effects.}

\item{iter}{\link{numeric}: Maximal number of iterations of the algorithm. The default is set to \code{iter=100}}

\item{time}{\link{vector}: Is the vector of the measurement times associated with the trajectories in \code{Y},\code{Z} and \code{X}.}

\item{sto}{\link{character}: Defines the covariance function of the stochastic process, can be either \code{"none"} for no stochastic process, \code{"BM"} for Brownian motion, \code{OrnUhl} for standard Ornstein-Uhlenbeck process, \code{BBridge} for Brownian Bridge, \code{fbm} for Fractional Brownian motion; can also be a function defined by the user.}

\item{delta}{\link{numeric}: The algorithm stops when the difference in log likelihood between two iterations is smaller than \code{delta}. The default value is set to O.O01}
}
\value{
A fitted (S)MERF model which is a list of the following elements: \itemize{
\item \code{forest:} Tree obtained at the last iteration.
\item \code{random_effects :} Predictions of random effects for different trajectories.
\item \code{id_btilde:} Identifiers of individuals associated with the predictions \code{random_effects}.
\item \code{var_random_effects: } Estimation of the variance covariance matrix of random effects.
\item \code{sigma_sto: } Estimation of the volatility parameter of the stochastic process.
\item \code{sigma: } Estimation of the residual variance parameter.
\item \code{time: } The vector of the measurement times associated with the trajectories in \code{Y},\code{Z} and \code{X}.
\item \code{sto: } Stochastic process used in the model.
\item \code{Vraisemblance:} Log-likelihood of the different iterations.
\item \code{id: } Vector of the identifiers for the different trajectories.
}
}
\description{
(S)MERT algorithm
}
\examples{
\dontrun{
set.seed(123)
data <- DataLongGenerator(n=20) # Generate the data composed by n=20 individuals.
# Train a SMERF model on the generated data. Should take ~ 50 secondes
# The data are generated with a Brownian motion,
# so we use the parameter sto="BM" to specify a Brownian motion as stochastic process
smert <- MERF(X=data$X,Y=data$Y,Z=data$Z,id=data$id,time=data$time,sto="BM")
smert$forest # is the fitted random forest (obtaind at the last iteration).
smert$random_effects are the predicted random effects for each individual.
smert$omega are the predicted stochastic processes.
plot(smerf$Vraisemblance) #evolution of the log-likelihood.
}

}
