% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongituRF.R
\name{Moy_fbm}
\alias{Moy_fbm}
\title{Title}
\usage{
Moy_fbm(id, Btilde, sigmahat, Phi, Y, Z, H, time, sigma2)

Moy_fbm(id, Btilde, sigmahat, Phi, Y, Z, H, time, sigma2)
}
\description{
Title

Title
}
\keyword{internal}
