% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongituRF.R
\name{opti.FBM}
\alias{opti.FBM}
\title{Title}
\usage{
opti.FBM(X, Y, id, Z, iter, mtry, ntree, time, conditional)
}
\arguments{
\item{X}{[matrix]}

\item{Y}{[vector]}

\item{id}{[vector]}

\item{Z}{[matrix]}

\item{iter}{[numeric]}

\item{mtry}{[numeric]}

\item{ntree}{[numeric]}

\item{time}{[time]}

\item{conditional}{[logical]}
}
\description{
Title
}
\keyword{internal}
