% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongituRF.R
\name{opti.FBMreem}
\alias{opti.FBMreem}
\title{Title}
\usage{
opti.FBMreem(X, Y, id, Z, iter, mtry, ntree, time)
}
\arguments{
\item{X}{[matrix]}

\item{Y}{[vector]}

\item{id}{[vector]}

\item{Z}{[matrix]}

\item{iter}{[numeric]}

\item{mtry}{[numeric]}

\item{ntree}{[numeric]}

\item{time}{[time]}
}
\description{
Title
}
\keyword{internal}
