% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongituRF.R
\name{predict.merf}
\alias{predict.merf}
\title{Predict with a (S)MERF alorithm}
\usage{
\method{predict}{merf}(merf_sto, X, Z, id, time)
}
\arguments{
\item{merf_sto}{:}

\item{X}{:}

\item{Z}{:}

\item{id}{:}

\item{time:}{}
}
\value{

}
\description{
Predict with a (S)MERF alorithm
}
