% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongituRF.R
\name{predict_lin}
\alias{predict_lin}
\title{Predict a fitted (S)LMEM algorithm}
\usage{
predict_lin(lin, X, Z, id, time)
}
\arguments{
\item{lin}{:}

\item{X}{:}

\item{Z}{:}

\item{id}{:}

\item{time}{:}
}
\value{

}
\description{
Predict a fitted (S)LMEM algorithm
}
