% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongituRF.R
\name{REEMforest}
\alias{REEMforest}
\title{(S)REEMforest alogorithm}
\usage{
REEMforest(X, Y, id, Z, iter, mtry, ntree, time, sto, delta = 0.001)
}
\arguments{
\item{X}{:}

\item{Y}{:}

\item{id}{:}

\item{Z}{:}

\item{iter}{:}

\item{mtry}{:}

\item{ntree}{:}

\item{time}{:}

\item{sto}{:}

\item{delta}{:}
}
\value{

}
\description{
(S)REEMforest alogorithm
}
