% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongituRF.R
\name{Stability_Score}
\alias{Stability_Score}
\title{Stability score function for (S)MERF and (S)REEMforest methods}
\usage{
Stability_Score(
  X,
  Y,
  Z,
  id,
  time,
  mtry,
  ntree,
  sto = "BM",
  method = "MERF",
  eta = c(1:ncol(X)),
  nvars = c(1:ncol(X))
)
}
\arguments{
\item{X}{[matrix]: A \code{N}x\code{p} matrix containing the \code{p} predictors of the fixed effects, column codes for a predictor.}

\item{Y}{[vector]: A vector containing the output trajectories.}

\item{Z}{[matrix]: A \code{N}x\code{q} matrix containing the \code{q} predictor of the random effects.}

\item{id}{[vector]: Is the vector of the identifiers for the different trajectories.}

\item{time}{[vector]: Is the vector of the measurement times associated with the trajectories in \code{Y},\code{Z} and \code{X}.}

\item{mtry}{[numeric]: Number of variables ramdomly picked to split each node.}

\item{ntree}{[numeric]: Number of trees in the RF.}

\item{sto}{[string]: Defines the covariance function of the stochastic process, can be either \code{"none"} for no stochastic process, \code{"BM"} for Brownian motion, \code{OrnUhl} for standard Ornstein-Uhlenbeck process, \code{BBridge} for Brownian Bridge, \code{fbm} for Fractional Brownian motion; can also be a function defined by the user.}

\item{method}{[string]: Defines the method to be used, can be either "MERF" or "REEMforest".}

\item{eta}{[numeric]: The size of the neighborhood for the stability score. Can be a vector, in this case, returns the stability scores corresponding to all the values of the vector.}

\item{nvars}{[numeric]: The number of variables to consider among the most impotant variables. Can be a vector, in this case, the function returns the stability scores corresponding to all the values of the vector.}
}
\value{
A matrix with all the stability scores corresponding to the eta and nvars values. The $i$th row corresponds to the $i$th value of eta while the $i$th column corresponds to the $i$ value of nvars.
}
\description{
Computes the stability scores for (S)MERF and (S)REEMforest methods.
}
