% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LongituRF.R
\name{MERT}
\alias{MERT}
\title{(S)MERT algorithm}
\usage{
MERT(X, Y, id, Z, iter, time, sto, delta = 0.001)
}
\arguments{
\item{X}{:}

\item{Y}{:}

\item{id}{:}

\item{Z}{:}

\item{iter}{:}

\item{time}{:}

\item{sto}{:}

\item{delta}{:}
}
\value{

}
\description{
(S)MERT algorithm
}
